backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.6

# AI/ML Dependencies
google-generativeai==0.3.2
google-api-python-client==2.95.0

# Text-to-Speech
pyttsx3==2.90

# Speech Recognition (optional)
SpeechRecognition==3.10.0

# Existing Flask dependencies (keep for compatibility)
Flask==2.3.3
flask-cors==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.8.0
requests==2.31.0
Werkzeug==2.3.7
Jinja2==3.1.2
bcrypt==4.0.1
